# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API de Contigo para acompañamiento de adultos mayores.
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  CLIENTE_DECISOR: 'CLIENTE_DECISOR',
  PERSONAL_ACOMPANANTE: 'PERSONAL_ACOMPANANTE',
  ADMINISTRADOR: 'ADMINISTRADOR',
} as const;

export interface User {
  id: string;
  email: string;
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dni?: string | null;
  role: UserRole;
}

export interface AuthSession {
  token: string;
  user: User;
}

export interface LoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface ClientRegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  fullName: string;
  /** @nullable */
  phone?: string | null;
  /**
     * @nullable
     * @pattern ^[0-9]{8}$
     */
  dni?: string | null;
}

export interface WorkerInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  fullName: string;
  phone: string;
  /** @pattern ^[0-9]{8}$ */
  dni: string;
  /** @nullable */
  skills?: string | null;
  backgroundChecked?: boolean;
}

export interface Senior {
  id: string;
  fullName: string;
  age: number;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
}

export interface SeniorInput {
  /** @minLength 2 */
  fullName: string;
  /**
     * @minimum 60
     * @maximum 130
     */
  age: number;
  /** @nullable */
  medicalNotes?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
}

export type WorkerStatus = typeof WorkerStatus[keyof typeof WorkerStatus];


export const WorkerStatus = {
  DISPONIBLE: 'DISPONIBLE',
  EN_SERVICIO: 'EN_SERVICIO',
  INACTIVO: 'INACTIVO',
} as const;

export interface Worker {
  id: string;
  fullName: string;
  email: string;
  phone: string;
  dni: string;
  status: WorkerStatus;
  /** @nullable */
  skills: string | null;
  backgroundChecked?: boolean;
  rating: number;
  completedServices?: number;
}

export interface ServiceOption {
  id: string;
  label: string;
  description: string;
}

export type RequestInputServiceType = typeof RequestInputServiceType[keyof typeof RequestInputServiceType];


export const RequestInputServiceType = {
  CITA_MEDICA: 'CITA_MEDICA',
  TRAMITE: 'TRAMITE',
  COMPRAS: 'COMPRAS',
  TECNOLOGIA: 'TECNOLOGIA',
  RECREACION: 'RECREACION',
} as const;

export interface RequestInput {
  seniorId: string;
  serviceType: RequestInputServiceType;
  scheduledAt: string;
  /**
     * @minimum 0.5
     * @maximum 24
     */
  durationHours: number;
  /** @minLength 3 */
  pickupAddress: string;
  /** @nullable */
  destinationAddress?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type StatusInputStatus = typeof StatusInputStatus[keyof typeof StatusInputStatus];


export const StatusInputStatus = {
  SOLICITADO: 'SOLICITADO',
  ASIGNADO: 'ASIGNADO',
  EN_CURSO: 'EN_CURSO',
  FINALIZADO: 'FINALIZADO',
  CANCELADO: 'CANCELADO',
} as const;

export interface StatusInput {
  status: StatusInputStatus;
}

export interface AssignmentInput {
  workerId: string;
}

export interface ServiceRequest {
  id: string;
  serviceType: string;
  serviceLabel: string;
  status: string;
  scheduledAt: string;
  durationHours: number;
  pickupAddress: string;
  /** @nullable */
  destinationAddress?: string | null;
  /** @nullable */
  notes?: string | null;
  clientName: string;
  seniorName: string;
  /** @nullable */
  workerId?: string | null;
  /** @nullable */
  workerName: string | null;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating: number | null;
  /** @nullable */
  reviewComment?: string | null;
  createdAt?: string;
}

export interface ReviewInput {
  requestId: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export interface Review {
  id: string;
  requestId: string;
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export type DashboardStats = {
  totalRequests: number;
  pendingRequests: number;
  activeServices: number;
  averageRating: number;
};

export interface Dashboard {
  role: string;
  stats: DashboardStats;
  recentRequests: ServiceRequest[];
}