# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API de Contigo para acompañamiento de adultos mayores.
 * OpenAPI spec version: 1.0.0
 */
import type { DashboardStats } from './dashboardStats';
import type { ServiceRequest } from './serviceRequest';

export interface Dashboard {
  role: string;
  stats: DashboardStats;
  recentRequests: ServiceRequest[];
}