# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API de Contigo para acompañamiento de adultos mayores.
 * OpenAPI spec version: 1.0.0
 */

export interface HealthStatus {
  status: string;
}