# lib/api-zod/src/generated/types/requestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API de Contigo para acompañamiento de adultos mayores.
 * OpenAPI spec version: 1.0.0
 */
import type { RequestInputServiceType } from './requestInputServiceType';

export interface RequestInput {
  seniorId: string;
  serviceType: RequestInputServiceType;
  scheduledAt: Date;
  /**
     * @minimum 0.5
     * @maximum 24
     */
  durationHours: number;
  /** @minLength 3 */
  pickupAddress: string;
  /** @nullable */
  destinationAddress?: string | null;
  /** @nullable */
  notes?: string | null;
}